Move rehasher to merkle package
