+++ merkle/log_proofs_test.go
@@ -18,8 +18,13 @@ import (
 	"fmt"
 	"testing"
 
+	//nolint:staticcheck
+	"github.com/golang/protobuf/proto"
 	"github.com/google/go-cmp/cmp"
+	"github.com/google/trillian"
 	"github.com/google/trillian/merkle/compact"
+	"github.com/google/trillian/merkle/rfc6962"
+	"github.com/google/trillian/storage/tree"
 )
 
 // TestCalcInclusionProofNodeAddresses contains inclusion proof tests. For
@@ -290,6 +295,82 @@ func TestConsistencySucceedsUpToTreeSize(t *testing.T) {
 				t.Errorf("CalcConsistencyProofNodeAddresses(%d, %d) = %v", s1, s2, err)
 			}
 		}
+	}
+}
+
+func TestRehasher(t *testing.T) {
+	th := rfc6962.DefaultHasher
+	h := [][]byte{
+		th.HashLeaf([]byte("Hash 1")),
+		th.HashLeaf([]byte("Hash 2")),
+		th.HashLeaf([]byte("Hash 3")),
+		th.HashLeaf([]byte("Hash 4")),
+		th.HashLeaf([]byte("Hash 5")),
+	}
+
+	for _, tc := range []struct {
+		desc   string
+		index  int64
+		hashes [][]byte
+		rehash []bool
+		want   [][]byte
+	}{
+		{
+			desc:   "no rehash",
+			index:  126,
+			hashes: h[:3],
+			rehash: []bool{false, false, false},
+			want:   h[:3],
+		},
+		{
+			desc:   "single rehash",
+			index:  999,
+			hashes: h[:5],
+			rehash: []bool{false, true, true, false, false},
+			want:   [][]byte{h[0], th.HashChildren(h[2], h[1]), h[3], h[4]},
+		},
+		{
+			desc:   "single rehash at end",
+			index:  11,
+			hashes: h[:3],
+			rehash: []bool{false, true, true},
+			want:   [][]byte{h[0], th.HashChildren(h[2], h[1])},
+		},
+		{
+			desc:   "single rehash multiple nodes",
+			index:  23,
+			hashes: h[:5],
+			rehash: []bool{false, true, true, true, false},
+			want:   [][]byte{h[0], th.HashChildren(h[3], th.HashChildren(h[2], h[1])), h[4]},
+		},
+		{
+			// TODO(pavelkalinnikov): This will never happen in our use-case. Design
+			// the type to not allow multi-rehash by design.
+			desc:   "multiple rehash",
+			index:  45,
+			hashes: h[:5],
+			rehash: []bool{true, true, false, true, true},
+			want:   [][]byte{th.HashChildren(h[1], h[0]), h[2], th.HashChildren(h[4], h[3])},
+		},
+	} {
+		t.Run(tc.desc, func(t *testing.T) {
+			r := Rehasher{Hasher: th}
+			for i, hash := range tc.hashes {
+				// TODO(pavelkalinnikov): Pass the hash and rehash directly.
+				r.Process(tree.Node{Hash: hash}, NodeFetch{Rehash: tc.rehash[i]})
+			}
+			got, err := r.RehashedProof(tc.index)
+			if err != nil {
+				t.Fatalf("rehashedProof: %v", err)
+			}
+			want := &trillian.Proof{
+				LeafIndex: tc.index,
+				Hashes:    tc.want,
+			}
+			if !proto.Equal(got, want) {
+				t.Errorf("proofs mismatch:\ngot: %v\nwant: %v", got, want)
+			}
+		})
 	}
 }
